Exctract utils from `src/api/items/index.ts`
